// dashboard/src/api/models/uiEventType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */

export type UiEventType = typeof UiEventType[keyof typeof UiEventType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UiEventType = {
  runstatus: 'run.status',
  runsnapshot: 'run.snapshot',
  runmetrics: 'run.metrics',
  nodestate: 'node.state',
  nodestatus: 'node.status',
  noderesultsnapshot: 'node.result.snapshot',
  noderesultdelta: 'node.result.delta',
  nodeerror: 'node.error',
  artifactready: 'artifact.ready',
  artifactremoved: 'artifact.removed',
  commandack: 'command.ack',
  commanderror: 'command.error',
  workerheartbeat: 'worker.heartbeat',
  workerpackage: 'worker.package',
} as const;