// dashboard/src/api/models/uiEventEnvelope.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { UiEventEnvelopeVersion } from './uiEventEnvelopeVersion';
import type { UiEventType } from './uiEventType';
import type { UiEventScope } from './uiEventScope';
import type { UiEventEnvelopeMeta } from './uiEventEnvelopeMeta';
import type { UiEventPayload } from './uiEventPayload';

export interface UiEventEnvelope {
  version: UiEventEnvelopeVersion;
  id: string;
  type: UiEventType;
  occurredAt: string;
  scope: UiEventScope;
  /** @nullable */
  replayed?: boolean | null;
  /** @nullable */
  correlationId?: string | null;
  /** @nullable */
  meta?: UiEventEnvelopeMeta;
  data: UiEventPayload;
}