// dashboard/src/api/models/nodeResultSnapshotEvent.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { NodeResultSnapshotEventFormat } from './nodeResultSnapshotEventFormat';
import type { NodeResultSnapshotEventContent } from './nodeResultSnapshotEventContent';
import type { RunArtifact } from './runArtifact';

export interface NodeResultSnapshotEvent {
  kind: 'node.result.snapshot';
  runId: string;
  nodeId: string;
  /** @minimum 0 */
  revision: number;
  format?: NodeResultSnapshotEventFormat;
  content: NodeResultSnapshotEventContent;
  /** @nullable */
  artifacts?: RunArtifact[] | null;
  /** @nullable */
  summary?: string | null;
  complete?: boolean;
}